2021 Day 6 - part 2 (credits: @mar-chi-pan)
The part 1 approach wouldn't scale, so this changes how things are done:

Instead of operating on each fish individually, we group all the fish by
their current day and we keep them in a map where the keys are the days,
and the values are the fish counts.

The map has a size of 9, i.e. indexes in the range 0-8.
We keep track of 2 pointers on this map: day0 and day6.

Day0 is where new fish are born and in day6 we add the fish that
just gave birth.

We don't need to care about day8, because that's automatically being
taken care off by the size of our map.